fix(data): Repair data in AuditLogEntry.data that shouldn't be there
We have two types of objects in `AuditLogEntry.data` that shouldn't be there:
- Generally, there are some rows that have a `BitHandler` in the `flags` key, when it should
  actually be an int. There's no specific error associated with this, but since we're iterating
  the table it seemed to make sense to fix
- For logs of type MEMBER_ACCEPT, there are some rows that have actual `Team` instances pickled in
  the `teams` key, rather than a list of ints, and a corresponding list of slugs in `teams_slugs`.
  Correcting this data fixes SENTRY-FAM

I've marked this as `is_dangerous` since we have low 10MM of rows in this table. Not sure how many
rows it'll update since there might be a lot of rows with that incorrect flags key. There are ~350k
max MEMBER_ACCEPT rows, so there shouldn't be too many updates there.

Tested this manually against an invalid pickle and things look to work as expected.

Once this has run, we can also remove the fixer code in
`sentry.api.serializers.models.auditlogentry.fix`. That should be in a separate pr though.